requirements.txt
django
django-geoposition
googlemaps
djangorestframework
dateutils
